♻️ (ical): Extract semantic Time type from typed ValueTime
- Add new semantic Time type in property_datetime.rs, update DateTime, Period
  to use semantic Time instead of ValueTime
- use u8 instead of i8 for hour/minute/second in ValueTime
- Make ValueUtcOffset.second Option<u8> instead of i8
- Rename Period::timezone() to jiff_timezone() for clarity

+++ ical/src/semantic/vtimezone.rs
@@ -353,10 +353,8 @@ impl TimeZoneOffset {
         match value {
             Value::UtcOffset(offset) => Ok(TimeZoneOffset {
                 positive: offset.positive,
-                #[allow(clippy::cast_sign_loss)]
-                hours: offset.hour as u8,
-                #[allow(clippy::cast_sign_loss)]
-                minutes: offset.minute as u8,
+                hours: offset.hour,
+                minutes: offset.minute,
             }),
             _ => Err(SemanticError::InvalidValue {
                 property: kind,
@@ -373,10 +371,8 @@ impl TryFrom<Value<'_>> for TimeZoneOffset {
         match value {
             Value::UtcOffset(offset) => Ok(TimeZoneOffset {
                 positive: offset.positive,
-                #[allow(clippy::cast_sign_loss)]
-                hours: offset.hour as u8,
-                #[allow(clippy::cast_sign_loss)]
-                minutes: offset.minute as u8,
+                hours: offset.hour,
+                minutes: offset.minute,
             }),
             _ => Err(SemanticError::InvalidValue {
                 property: PropertyKind::TzOffsetFrom, // Default fallback
